requirements.txt
streamlit==1.49.0
python-dotenv

llama-index==0.12.0
llama-index-core==0.12.52.post1
llama-index-legacy==0.9.48.post4
llama-index-instrumentation==0.4.2
llama-index-llms-groq==0.3.1
llama-index-embeddings-huggingface==0.5.5
llama-index-readers-file==0.4.11
llama-index-readers-llama-parse==0.4.0

faiss-cpu==1.13.1

huggingface-hub==0.34.4
sentence-transformers==4.1.0

pymupdf==1.26.4
pypdf==5.9.0

json
regex==2025.7.34
